requirements.txt
requests
aiohttp
discord.py
Pillow
PyMySQL